[refactor] S1: restructure packages — delete design/, move pkg/io to io/, rename internal/transport to internal/engine, inline internal/errors
- Delete design/ directory (draft files should not be in public module)
- Move pkg/io/ to io/ and update all imports across 20+ files
- Rename internal/transport/ to internal/engine/ to avoid naming conflict with public transport/ package
- Inline internal/errors into fmt.Errorf and delete the package
