+++ codec/packet/delimiter_test.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"github.com/emove/less/codec"
 	"github.com/emove/less/codec/payload"
-	"github.com/emove/less/pkg/io"
-	ior "github.com/emove/less/pkg/io/reader"
-	"github.com/emove/less/pkg/io/writer"
+	"github.com/emove/less/io"
+	ior "github.com/emove/less/io/reader"
+	"github.com/emove/less/io/writer"
 	"reflect"
 	"testing"
 )
