+++ codec/payload/json_test.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/emove/less/codec"
-	"github.com/emove/less/pkg/io"
-	"github.com/emove/less/pkg/io/reader"
-	"github.com/emove/less/pkg/io/writer"
+	"github.com/emove/less/io"
+	"github.com/emove/less/io/reader"
+	"github.com/emove/less/io/writer"
 	"reflect"
 	"testing"
 )
