+++ internal/engine/options.go
@@ -1,4 +1,4 @@
-package transport
+package engine
 
 import (
 	"math"
