--- internal/errors/errors.go
@@ -1,7 +1,0 @@
-package errors
-
-import "fmt"
-
-func New(format string, args ...interface{}) error {
-	return fmt.Errorf(format, args...)
-}
