+++ internal/recovery/recover.go
@@ -1,14 +1,14 @@
 package recovery
 
 import (
-	"github.com/emove/less/internal/errors"
+	"fmt"
 	"runtime/debug"
 )
 
 func Do(fn func() error) (err error) {
 	defer func() {
 		if p := recover(); p != nil {
-			err = errors.New("panic: %v\n stack: %s", p, string(debug.Stack()))
+			err = fmt.Errorf("panic: %v\n stack: %s", p, string(debug.Stack()))
 		}
 	}()
 
@@ -17,7 +17,7 @@ func Do(fn func() error) (err error) {
 
 func Recover(fn func(err error)) {
 	if p := recover(); p != nil {
-		err := errors.New("panic error: %v\n stack: %s", p, string(debug.Stack()))
+		err := fmt.Errorf("panic error: %v\n stack: %s", p, string(debug.Stack()))
 		fn(err)
 	}
 }
